Updated to code released with FaceFX 2015. Changes include:
Frame 0 is now used for the bind pose by default to help compute a correct bounding box. Binding boxes are also computed from mesh vertices instead of bones influencing the mesh for more accurate boxes.

Fixed bug where FxOgreFBX could potentially create an invalid MESH file that referenced Ogre poses that did not exist causing Ogre to crash.
Negative time keys are no longer supported in FBX files, resolving an issue where ogre animations are shifted forward in time.

Fixed memory leak in FxOgreFBX where delete was being called on an array of Mesh vertices instead of delete[]

updated to the 2015.1 fbx sdk.

Fixed potential issue with FBX morph target export related to the bind position.

Fixed potential morph-related crash in ExecOgreFBX that was occuring with DAZ-generated FBX files when using the FBX bind information.

Fixed blendshape animations for non-shared geometry in FxOgreFBX. Removed some unused and minimally used code to clear up morph target codepaths.

The FBX importer will now only take the first 8 texture coordinates if there are more than 8 texture coordinates, to preserve compatibility with Direct3D.

Added framework for displaying a warning dialog for FBX import errors. Too many bones, duplicate named bones, file format errors, textures not found, unsupported UVs, no meshes exported will generate warnings of this type.

Duplicate named joints no longer fails the export

Fixed crash exporting non-morph content without shared geometry

Improved support for point/omni lights in FxOgreFBX

The FBX exporter no longer exports blendshape animations for .MESH files with no blendshapes

Fixed case when no bounding box is computed because the skeleton does not exist or does not yield an acceptable box.

fixed several issues with calculating the bounding box and the resulting scene transform scale that is calculated from the bounding box. A mesh can no longer be so large that it appears behind the far clipping plane.

upgraded Ogre FBX exporter to tinyxml2

Fixed transparency issue on materials that have no associated texture file.

Fixed 2 texture-path related issues. FileExists would return true when passed an invalid path. Also relative textures would only work if the texture path was already relative...now it works if the original texture path was a fully qualified path that   does not exist, but a texture with the same name exists in the FBX directory.

+++ Src/paramlist.cpp
@@ -34,8 +34,8 @@ namespace FxOgreFBX
     // Helper function for getting the filename from a full path.
     std::string StripToTopParent(const std::string& filepath)
     {
-        int ri = filepath.find_last_of('\\');
-        int ri2 = filepath.find_last_of('/');
+        int ri = static_cast<int>(filepath.find_last_of('\\'));
+        int ri2 = static_cast<int>(filepath.find_last_of('/'));
         if(ri2 > ri)
         {
             ri = ri2;
@@ -45,8 +45,8 @@ namespace FxOgreFBX
 
     std::string StripFilename(const std::string& filepath)
     {
-        int ri = filepath.find_last_of('\\');
-        int ri2 = filepath.find_last_of('/');
+        int ri = static_cast<int>(filepath.find_last_of('\\'));
+        int ri2 = static_cast<int>(filepath.find_last_of('/'));
         if(ri2 > ri)
         {
             ri = ri2;
